fixed/improved docu, renamed "machine_learning" into "forecasting" Co-authored-by: Felipe Trost <felipe.trost@gmail.com>
