--- docs/sdk/code-reference/pipelines/data_quality/data_manipulation/spark/prediction/auto_arima.md
@@ -1,1 +1,0 @@
-::: src.sdk.python.rtdip_sdk.pipelines.data_quality.data_manipulation.spark.prediction.auto_arima
