+++ docs/sdk/code-reference/pipelines/forecasting/spark/arima.md
@@ -1,0 +1,1 @@
+::: src.sdk.python.rtdip_sdk.pipelines.forecasting.spark.arima
