+++ docs/sdk/code-reference/pipelines/forecasting/spark/data_binning.md
@@ -1,0 +1,1 @@
+::: src.sdk.python.rtdip_sdk.pipelines.forecasting.spark.data_binning
