--- docs/sdk/code-reference/pipelines/machine_learning/spark/linear_regression.md
@@ -1,1 +1,0 @@
-::: src.sdk.python.rtdip_sdk.pipelines.machine_learning.spark.linear_regression
