+++ src/sdk/python/rtdip_sdk/pipelines/data_quality/data_manipulation/spark/__init__.py
@@ -21,3 +21,4 @@ from .k_sigma_anomaly_detection import KSigmaAnomalyDetection
 from .missing_value_imputation import MissingValueImputation
 from .out_of_range_value_filter import OutOfRangeValueFilter
 from .flatline_filter import FlatlineFilter
+from .missing_value_imputation import MissingValueImputation
