+++ src/sdk/python/rtdip_sdk/pipelines/data_quality/data_manipulation/spark/normalization/denormalization.py
@@ -27,7 +27,6 @@ from .normalization import (
 
 class Denormalization(DataManipulationBaseInterface, InputValidator):
     """
-    #TODO
     Applies the appropriate denormalization method to revert values to their original scale.
 
     Example
